// lib/netlify/schema.ts
export interface components {
    schemas: {
// ...
        site: {
            id?: string
            state?: string
            plan?: string
            name?: string
            custom_domain?: string
            domain_aliases?: string[]
            branch_deploy_custom_domain?: string
            deploy_preview_custom_domain?: string
            password?: string
            notification_email?: string
            url?: string
            ssl_url?: string
            admin_url?: string
            screenshot_url?: string
            /** Format: dateTime */
            created_at?: string
            /** Format: dateTime */
            updated_at?: string
            user_id?: string
            session_id?: string
            ssl?: boolean
            force_ssl?: boolean
            managed_dns?: boolean
            deploy_url?: string
            published_deploy?: components['schemas']['deploy']
            account_id?: string
            account_name?: string
            account_slug?: string
            git_provider?: string
            deploy_hook?: string
            capabilities?: {
                [key: string]: Record<string, never>
            }
            processing_settings?: {
                html?: {
                    pretty_urls?: boolean
                }
                ignore_html_forms?: boolean
            }
            build_settings?: components['schemas']['repoInfo']
            id_domain?: string
            default_hooks_data?: {
                access_token?: string
            }
            build_image?: string
            prerender?: string
            functions_region?: string
        }
// ...
        deploy: {
            id?: string
            site_id?: string
            user_id?: string
            build_id?: string
            state?:
                | 'new'
                | 'pending_review'
                | 'accepted'
                | 'rejected'
                | 'enqueued'
                | 'building'
                | 'uploading'
                | 'uploaded'
                | 'preparing'
                | 'prepared'
                | 'processing'
                | 'processed'
                | 'ready'
                | 'error'
                | 'retrying'
            name?: string
            url?: string
            ssl_url?: string
            admin_url?: string
            deploy_url?: string
            deploy_ssl_url?: string
            screenshot_url?: string
            review_id?: number
            draft?: boolean
            required?: string[]
            required_functions?: string[]
            error_message?: string
            branch?: string
            commit_ref?: string
            commit_url?: string
            skipped?: boolean
            /** Format: dateTime */
            created_at?: string
            /** Format: dateTime */
            updated_at?: string
            /** Format: dateTime */
            published_at?: string
            title?: string
            context?: string
            locked?: boolean
            review_url?: string
            framework?: string
            function_schedules?: components['schemas']['functionSchedule'][]
            commit_message?: string
            deploy_time?: number
            available_functions?: any[]
            committer?: string
            skipped_log?: string
            manual_deploy?: boolean
            plugin_state?: string
            lighthouse_plugin_scores?: Record<string, any>
            links: {
                permalink: string
                alias: string
                branch: null
            }
            entry_path?: string
            views_count?: any
            public_repo?: string
            pending_review_reason?: any
            lighthouse?: any
            edge_functions_present?: any
            expires_at?: string
            blobs_region?: string
            summary?: {
                status: 'ready' | string
                messages: {
                    type: 'info' | string
                    title: string
                    description: string
                    details: null | string
                }[]
            }
        }
// ...
    }
// ...
}